requirements.txt
python-dotenv
openai
black